extract 'create entity type' form, handle extension. add 'extend type' context menu item
